+++ Assets/3.Scripts/ScriptableObject/PlayerData.cs
@@ -12,6 +12,6 @@ public class PlayerData : ScriptableObject
     public float DamagePercent { get { return damagePercent; } set { damagePercent = value; } }
 
     [SerializeField]
-    private int hp;
-    public int HP { get { return hp; } }
+    private float hp;
+    public float HP { get { return hp; } }
 }
